+++ livekit/src/plugin.rs
@@ -1,10 +1,14 @@
 use std::{
     ffi::{c_char, c_void, CString},
+    pin::Pin,
     sync::Arc,
+    task::{Context, Poll},
+    time::Duration,
 };
 
+use futures_util::Stream;
 use libloading::{Library, Symbol};
-use parking_lot::{Mutex, RwLock};
+use libwebrtc::{audio_stream::native::NativeAudioStream, prelude::AudioFrame};
 
 #[derive(Debug, thiserror::Error)]
 pub enum PluginError {
@@ -96,6 +100,67 @@ impl Drop for AudioFilterSession {
         if !self.ptr.is_null() {
             self.destroy();
         }
+    }
+}
+
+pub struct AudioFilterAudioStream {
+    inner: NativeAudioStream,
+    session: AudioFilterSession,
+    buffer: Vec<i16>,
+    sample_rate: u32,
+    num_channels: u32,
+    frame_size: usize,
+}
+
+impl AudioFilterAudioStream {
+    pub fn new(
+        inner: NativeAudioStream,
+        session: AudioFilterSession,
+        duration: Duration,
+        sample_rate: u32,
+        num_channels: u32,
+    ) -> Self {
+        let frame_size =
+            ((sample_rate as f64) * duration.as_secs_f64() * num_channels as f64) as usize;
+        Self {
+            inner,
+            session,
+            buffer: Vec::with_capacity(frame_size),
+            sample_rate,
+            num_channels,
+            frame_size,
+        }
+    }
+}
+
+impl Stream for AudioFilterAudioStream {
+    type Item = AudioFrame<'static>;
+
+    fn poll_next(self: Pin<&mut Self>, cx: &mut Context) -> Poll<Option<Self::Item>> {
+        let this = self.get_mut();
+
+        while let Poll::Ready(frame) = Pin::new(&mut this.inner).poll_next(cx) {
+            let Some(frame) = frame else {
+                return Poll::Ready(None);
+            };
+            this.buffer.extend_from_slice(&frame.data);
+
+            if this.buffer.len() >= this.frame_size {
+                let data = this.buffer.drain(..this.frame_size).collect::<Vec<_>>();
+                let mut out: Vec<i16> = Vec::with_capacity(this.frame_size);
+
+                this.session.process_i16(this.frame_size, &data, &mut out);
+
+                return Poll::Ready(Some(AudioFrame {
+                    data: out.into(),
+                    sample_rate: this.sample_rate,
+                    num_channels: this.num_channels,
+                    samples_per_channel: (this.frame_size / this.num_channels as usize) as u32,
+                }));
+            }
+        }
+
+        Poll::Pending
     }
 }
